+++ gradle/libs.versions.toml
@@ -5,6 +5,7 @@ kotlin = "1.9.23"
 androidMinSdk = "24"
 androidCompileSdk = "34"
 androidTargetSdk = "34"
+buildKonfig = "0.15.1"
 kotlinxSerialization = "1.6.3"
 kotlinTestingResource = "0.4.1"
 ktlintVersion = "12.1.1"
@@ -93,8 +94,8 @@ kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref =
 kotlinPluginSerialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
 kotlinTestingResource = { id = "com.goncalossilva.resources", version.ref = "kotlinTestingResource" }
 ktlint = { id = "org.jlleitschuh.gradle.ktlint", version.ref = "ktlintVersion" }
+buildKonfig = { id = "com.codingfeline.buildkonfig", version.ref = "buildKonfig" }
 mavenPublish = { id = "maven-publish", version = "unspecified" }
 commonMppLib = { id = "common.mpp.lib", version = "unspecified" }
 commonMppPublish = { id = "common.mpp.publish", version = "unspecified" }
 commonMppCompose = { id = "common.mpp.compose", version = "unspecified" }
-commonMppBuildKonfig = { id = "common.mpp.buildKonfig", version = "unspecified" }
